No Failure if file is not accessable - skips if file is not readable or writeable - shows summary of skipped files at final success and stopped popup messageboxes
Improved Code and Error Messages
- if path does not exist: now has a clear message with the missing path on the right spot
- outsourced some code into seperate methods and cleaned some stuff up
- improved error message if user did not check any of the edit-options

+++ S3PR/s3molib/Package.cs
@@ -179,30 +179,65 @@ namespace s3molib
                 Package importPackage = source.ElementAt(i);
                 String originalPath = importPackage.FilePath;
                 String tempPath = Path.Combine(Path.GetDirectoryName(originalPath), Path.GetFileNameWithoutExtension(originalPath) + ".temp" + Path.GetExtension(originalPath));
-                Package package = Package.New(tempPath);
-                FileStream stream = new FileStream(tempPath, FileMode.Open, FileAccess.Write, FileShare.Read);
-                BinaryWriter w = new BinaryWriter(stream);
-                stream.Position = (long)((ulong)package.IndexPosition);
-                foreach (ResourceEntry re in importPackage.ResourceEntries)
-                {
-                    if (!(Helper.THUMResources.Contains(re.Type) && removeThumbnails == true) && !(Helper.ICONResources.Contains(re.Type) && removeIcons == true))
+                try {
+                    Package package = Package.New(tempPath);
+                    FileStream stream = new FileStream(tempPath, FileMode.Open, FileAccess.Write, FileShare.Read);
+                    BinaryWriter w = new BinaryWriter(stream);
+                    stream.Position = (long)((ulong)package.IndexPosition);
+                    foreach (ResourceEntry re in importPackage.ResourceEntries)
                     {
-                        byte[] data = importPackage.GetRawData(re);
-                        if (data == null)
+                        if (!(Helper.THUMResources.Contains(re.Type) && removeThumbnails == true) && !(Helper.ICONResources.Contains(re.Type) && removeIcons == true))
                         {
-                            throw new Exception("Unable to obtain data from " + Path.GetFileName(importPackage.FilePath));
+                            byte[] data = importPackage.GetRawData(re);
+                            if (data == null)
+                            {
+                                throw new Exception("Unable to obtain data from " + Path.GetFileName(importPackage.FilePath));
+                            }
+                            package._resourceEntries.Add(new ResourceEntry(importPackage, re.Type, re.Group, re.ID1, re.ID2, (uint)stream.Position, re.FileSize, re.MemSize, re.Compressed, re.Unknown2));
+                            w.Write(data, 0, (int)re.FileSize);
+                       }
+                    }
+                    package.IndexPosition = (uint)stream.Position;
+                    package.WriteIndex(w);
+                    package.WriteHeader(stream, w);
+                    w.Close();
+                    stream.Close();
+
+                    if (File.Exists(originalPath))
+                    {
+                        try
+                        {
+                            FileAttributes attrs = File.GetAttributes(originalPath);
+                            if ((attrs & FileAttributes.ReadOnly) == FileAttributes.ReadOnly)
+                            {
+                                File.SetAttributes(originalPath, attrs & ~FileAttributes.ReadOnly);
+                            }
                         }
-                        package._resourceEntries.Add(new ResourceEntry(importPackage, re.Type, re.Group, re.ID1, re.ID2, (uint)stream.Position, re.FileSize, re.MemSize, re.Compressed, re.Unknown2));
-                        w.Write(data, 0, (int)re.FileSize);
-                   }
+                        catch { }
+
+                        File.Delete(originalPath);
+                    }
+
+                    File.Move(tempPath, originalPath);
                 }
-                package.IndexPosition = (uint)stream.Position;
-                package.WriteIndex(w);
-                package.WriteHeader(stream, w);
-                w.Close();
-                stream.Close();
-                File.Delete(originalPath);
-                File.Move(tempPath, originalPath);
+                catch (UnauthorizedAccessException)
+                {
+                    // Skip files without permission
+                    continue;
+                }
+                catch (IOException)
+                {
+                    // Skip files that cannot be accessed or replaced
+                    continue;
+                }
+                finally
+                {
+                    // Remove temp files, if they still are there, just in case
+                    if (File.Exists(tempPath))
+                    {
+                        File.Delete(tempPath);
+                    }
+                }
             }
         }
